[nfc] remnants of includes cleanup
