+++ src/workerd/api/blob.h
@@ -6,7 +6,9 @@
 
 #include <workerd/io/compatibility-date.capnp.h>
 #include <workerd/jsg/jsg.h>
+#include <workerd/jsg/promise.h>
 namespace workerd { namespace api { class ReadableStream; } }
+using kj::byte;
 
 namespace workerd::api {
