+++ src/workerd/api/crypto.c++
@@ -3,11 +3,13 @@
 //     https://opensource.org/licenses/Apache-2.0
 
 #include "crypto.h"
+#include "crypto-streams.h"
 #include "crypto-impl.h"
 #include <array>
 #include <openssl/crypto.h>
 #include <openssl/err.h>
 #include <workerd/jsg/jsg.h>
+#include <workerd/jsg/value.h>
 #include "util.h"
 #include <workerd/io/io-context.h>
 #include <workerd/util/uuid.h>
