+++ src/workerd/api/form-data.h
@@ -7,9 +7,10 @@
 #include <kj/vector.h>
 #include <kj/compat/url.h>
 #include <workerd/jsg/jsg.h>
+#include <workerd/jsg/iterator.h>
 #include "blob.h"
 #include <workerd/io/compatibility-date.capnp.h>
-
+using kj::uint;
 namespace workerd::api {
 
 namespace url {
