+++ src/workerd/api/node/buffer.c++
@@ -10,6 +10,9 @@
 #include <kj/encoding.h>
 #include <algorithm>
 
+using kj::uint;
+using kj::byte;
+
 // These are defined by <sys/byteorder.h> or <netinet/in.h> on some systems.
 // To avoid warnings, undefine them before redefining them.
 #ifdef BSWAP_2
