+++ src/workerd/api/util.h
@@ -8,7 +8,6 @@
 #include <kj/compat/url.h>
 #include <kj/string.h>
 #include <workerd/jsg/jsg.h>
-#include <v8.h>
 
 namespace workerd::api {
