+++ src/workerd/io/actor-cache.c++
@@ -9,11 +9,12 @@
 
 #include <workerd/jsg/jsg.h>
 #include <workerd/io/io-gate.h>
+#include <workerd/io/io-context.h>
 #include <workerd/util/sentry.h>
 
 // TODO(cleanup): Break dependency on IoContext. ActorCache is intended to be independent of the
 //   rest of the codebase.
-//#include <workerd/io/io-context.h>
+#include <workerd/io/io-context.h>
 
 namespace workerd {
