+++ src/workerd/io/compatibility-date.c++
@@ -9,6 +9,7 @@
 #include <capnp/dynamic.h>
 #include <kj/map.h>
 
+
 namespace workerd {
 
 using kj::uint;
@@ -58,7 +59,7 @@ struct CompatDate {
     return CompatDate { year, month, day };
   }
 
-  static CompatDate parse(kj::StringPtr text, Worker::ValidationErrorReporter& errorReporter) {
+  static CompatDate parse(kj::StringPtr text, ValidationErrorReporter& errorReporter) {
     static constexpr CompatDate DEFAULT_DATE { 2021, 5, 1 };
     KJ_IF_MAYBE(v, parse(text)) {
       return *v;
@@ -89,7 +90,7 @@ struct CompatDate {
 
 void compileCompatibilityFlags(kj::StringPtr compatDate, capnp::List<capnp::Text>::Reader compatFlags,
                          CompatibilityFlags::Builder output,
-                         Worker::ValidationErrorReporter& errorReporter,
+                         ValidationErrorReporter& errorReporter,
                          bool allowExperimentalFeatures,
                          CompatibilityDateValidation dateValidation) {
   auto parsedCompatDate = CompatDate::parse(compatDate, errorReporter);
