+++ src/workerd/jsg/dom-exception.c++
@@ -3,6 +3,7 @@
 //     https://opensource.org/licenses/Apache-2.0
 
 #include "dom-exception.h"
+#include "resource.h"
 #include <kj/string.h>
 #include <map>
