+++ src/workerd/jsg/promise.h
@@ -7,7 +7,7 @@
 #include "util.h"
 #include "wrappable.h"
 #include "jsg.h"
-#include "web-idl.h"
+#include <kj/table.h>
 #include <kj/async.h>
 
 namespace workerd::jsg {
