+++ src/workerd/jsg/setup.h
@@ -6,8 +6,7 @@
 // Public API for setting up JavaScript context. Only high-level code needs to include this file.
 
 #include "jsg.h"
-#include "async-context.h"
-#include "type-wrapper.h"
+#include "function.h"
 #include "v8-platform-wrapper.h"
 #include <workerd/util/batch-queue.h>
 #include <kj/map.h>
