+++ src/workerd/server/workerd-api.h
@@ -29,7 +29,7 @@ public:
 
   static Worker::Script::Source extractSource(kj::StringPtr name,
       config::Worker::Reader conf,
-      Worker::ValidationErrorReporter& errorReporter,
+      ValidationErrorReporter& errorReporter,
       capnp::List<config::Extension>::Reader extensions);
 
   struct Global {
@@ -157,10 +157,10 @@ private:
 
   kj::Array<Worker::Script::CompiledGlobal> compileScriptGlobals(
       jsg::Lock& lock, config::Worker::Reader conf,
-      Worker::ValidationErrorReporter& errorReporter) const;
+      ValidationErrorReporter& errorReporter) const;
   kj::Own<jsg::ModuleRegistry> compileModules(
       jsg::Lock& lock, config::Worker::Reader conf,
-      Worker::ValidationErrorReporter& errorReporter,
+      ValidationErrorReporter& errorReporter,
       capnp::List<config::Extension>::Reader extensions) const;
 };
